Add: basic Rails installation
This is the result of starting with an empty directory and running:

  rails new .

(also could have been rails new score-bored from the parent folder)
this installed the latest version of Rails 7 and all of its
scaffolding. There is ALOT here which enables code features from the
JavaScript front end elements all the way to the SQL database backend.